# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UNIPROM CRM API
  version: 1.0.0
  description: |
    HTTP API CRM для обработки заявок на изготовление изделий.

    Два взаимозаменяемых префикса с одинаковой серверной проверкой прав и области
    видимости: канонический `/api/v1/...` и контрактный `/api/v1/crm/...`.

    Аутентификация — JWT в заголовке `Authorization: Bearer <token>`.
    Технические эндпоинты дополнительно требуют заголовок `x-integration-token`.

    Создание заявки принимает заголовок `Idempotency-Key`. Изменение ресурса
    требует `lockVersion`; при конфликте возвращается 409 `VERSION_CONFLICT`.
  contact:
    name: UNIPROM
servers:
  - url: http://127.0.0.1:3001
    description: Локальный backend
...
paths:
...
  /api/v1/applications/{number}/quotes:
    get:
      tags: [quotes]
      summary: Версии КП по заявке
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: Версии
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Quote' }
    post:
      tags: [quotes]
      summary: Создание версии КП
      parameters:
        - $ref: '#/components/parameters/Number'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                amount: { type: number }
                currency: { type: string }
                discountPct: { type: number }
                taxPct: { type: number }
                validUntil: { type: string, format: date-time }
                lines:
                  type: array
                  items:
                    type: object
                    properties:
                      lineId: { type: string }
                      name: { type: string }
                      quantity: { type: number }
                      price: { type: number }
      responses:
        '201': { description: Версия создана }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/quotes/{quoteId}:
    get:
      tags: [quotes]
      summary: Версия КП
      parameters:
        - $ref: '#/components/parameters/QuoteId'
      responses:
        '200':
          description: Версия КП
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Quote' }
    patch:
      tags: [quotes]
      summary: Изменение версии КП
      parameters:
        - $ref: '#/components/parameters/QuoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lockVersion]
              properties:
                amount: { type: number }
                currency: { type: string }
                validUntil: { type: string, format: date-time }
                lockVersion: { type: integer }
      responses:
        '200': { description: Версия изменена }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/quotes/{quoteId}/generate:
    post:
      tags: [quotes]
      summary: Генерация КП по расчёту
      parameters:
        - $ref: '#/components/parameters/QuoteId'
      responses:
        '200': { description: Версия сформирована }
  /api/v1/quotes/{quoteId}/upload:
    post:
      tags: [quotes]
      summary: Загрузка файла КП
      description: Загруженная и сгенерированная версии проходят одинаковые обязательные проверки.
      parameters:
        - $ref: '#/components/parameters/QuoteId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { description: Файл прикреплён, версия переведена в READY }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/quotes/{quoteId}/submit-approval:
    post:
      tags: [quotes]
      summary: Отправка КП на согласование
      description: Проверяется заключение КО для CUSTOM и для MODIFIED с обязательным участием КО.
      parameters:
        - $ref: '#/components/parameters/QuoteId'
      responses:
        '200': { description: Отправлено руководителю }
        '422':
          description: Требуется действующее положительное заключение КО
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/quotes/{quoteId}/approval-decision:
    post:
      tags: [quotes]
      summary: Решение руководителя по КП
      parameters:
        - $ref: '#/components/parameters/QuoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision, lockVersion]
              properties:
                decision: { type: string, enum: [APPROVED, REJECTED] }
                comment: { type: string, description: Обязателен при отклонении }
                lockVersion: { type: integer }
      responses:
        '200': { description: Решение зафиксировано, автор уведомлён }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Содержимое изменилось после согласования
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/quotes/{quoteId}/send:
    post:
      tags: [quotes]
      summary: Отправка КП клиенту
      parameters:
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': { description: Создано одно почтовое задание }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/v1/quotes/{quoteId}/customer-decision:
    post:
      tags: [quotes]
      summary: Решение клиента по КП
      parameters:
        - $ref: '#/components/parameters/QuoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [ACCEPTED, DECLINED] }
                comment: { type: string }
      responses:
        '200': { description: Решение клиента зафиксировано }
  /api/v1/quotes/{quoteId}/external-send:
    post:
      tags: [quotes]
      summary: Фиксация внешней отправки
      description: Фиксируются версия, получатель, фактическое время и ручной источник.
      parameters:
        - $ref: '#/components/parameters/QuoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipients, sentAt]
              properties:
                recipients:
                  type: array
                  items: { type: string }
                sentAt: { type: string, format: date-time }
                channel: { type: string }
      responses:
        '200': { description: Отправка зафиксирована }
  /api/v1/applications/{number}/commercial-basis-check:
    get:
      tags: [quotes]
      summary: Проверка коммерческого основания
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: Результат проверки
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckResult' }
...
  /api/v1/applications/{number}/fulfilment:
    get:
      tags: [production]
      summary: Состояние исполнения
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: Отгрузки, оплаты, остатки по позициям
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Fulfilment' }
  /api/v1/applications/{number}/close-check:
    get:
      tags: [production]
      summary: Проверка возможности закрытия
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: Блокирующие условия закрытия
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckResult' }
  /api/v1/applications/{number}/close:
    post:
      tags: [production]
      summary: Закрытие заявки
      description: |
        Отдельное действие. Блокируют неполная отгрузка, незавершённая постоплата
        и отсутствие обязательных документов. При успехе фиксируются время и автор
        закрытия, событие попадает в историю.
      parameters:
        - $ref: '#/components/parameters/Number'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lockVersion]
              properties:
                comment: { type: string }
                lockVersion: { type: integer }
      responses:
        '200': { description: Заявка закрыта }
        '409':
          description: Условия закрытия не выполнены
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckResult' }
  /api/v1/applications/{number}/closing-documents:
    get:
      tags: [production]
      summary: Обязательные закрывающие документы
      parameters:
        - $ref: '#/components/parameters/Number'
      responses:
        '200':
          description: Документы и их состояние
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ClosingDocument' }
    post:
      tags: [production]
      summary: Регистрация закрывающего документа
      parameters:
        - $ref: '#/components/parameters/Number'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [documentType]
              properties:
                documentType: { type: string }
                documentNumber: { type: string }
                status: { type: string, enum: [REGISTERED, SIGNED] }
      responses:
        '201': { description: Документ зарегистрирован }
...
  /api/v1/sla/instances/{instanceId}/resume:
    post:
      tags: [admin]
      summary: Возобновление приостановленного SLA
      parameters:
        - name: instanceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Отсчёт возобновлён }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    integrationToken:
      type: apiKey
      in: header
      name: x-integration-token
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: Повтор с тем же ключом не создаёт вторую запись.
    Number:
      name: number
      in: path
      required: true
      schema: { type: string }
      description: Номер заявки.
...
    QuoteId:
      name: quoteId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Ошибка запроса
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Конфликт состояния или версии
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Нарушено бизнес-правило
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message, correlationId]
      properties:
        code: { type: string, example: VERSION_CONFLICT }
        message: { type: string }
        correlationId: { type: string }
        fields:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    Quote:
      type: object
      properties:
        id: { type: string }
        versionNo: { type: integer }
        source: { type: string, enum: [GENERATED, UPLOADED] }
        status: { type: string }
        approvalStatus: { type: string }
        amount: { type: number, nullable: true }
        currency: { type: string, nullable: true }
        validUntil: { type: string, format: date-time, nullable: true }
        approvalFingerprint: { type: string, nullable: true }
        lockVersion: { type: integer }
        lines:
          type: array
          items:
            type: object
            properties:
              lineId: { type: string }
              name: { type: string }
              quantity: { type: number }
              price: { type: number }
    CheckResult:
      type: object
      properties:
        ok: { type: boolean }
        blockers:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
        detail:
          type: object
          additionalProperties: true
...
    Fulfilment:
      type: object
      properties:
        shippedQuantity: { type: number }
        remainingQuantity: { type: number }
        paidAmount: { type: number }
        openInvoices: { type: integer }
        lines:
          type: array
          items:
            type: object
            properties:
              lineId: { type: string }
              ordered: { type: number }
              shipped: { type: number }
              remaining: { type: number }
    ClosingDocument:
      type: object
      properties:
        id: { type: string }
        documentType: { type: string }
        documentNumber: { type: string, nullable: true }
        status: { type: string, enum: [REGISTERED, SIGNED] }
...
security:
  - bearerAuth: []